# Improve performance of Angle.wrap_at