@maxnoe 2022-07-28
Benchmark:

```
import astropy.units as u
from astropy.units import Quantity
import numpy as np
from astropy.coordinates.angles import Longitude, Latitude, Angle


rng = np.random.default_rng(0)

N = (1, 100, 10000)

p_nan = (0, 1e-2, 0.1, 0.5)
wrap = Angle(180, u.deg)


for n in N:
    for p in p_nan:
        values = rng.uniform(-720, 720, n)
        if p > 0:
            values[rng.choice(n, int(np.ceil(p * n)), replace=False)] = np.nan

        angle = Angle(values, u.deg)

        print('total', n, 'nans', np.count_nonzero(np.isnan(values)))
        get_ipython().run_line_magic('timeit', f'angle.wrap_at(wrap)')
```

Results:

```
This Branch:          Master:
total 1 nans 0
42.4 µs ± 798 ns      59.2 µs ± 839 ns
total 1 nans 1
32.5 µs ± 107 ns      48.8 µs ± 440 ns
total 1 nans 1
32.2 µs ± 89.6 ns     48.5 µs ± 56.4 ns
total 1 nans 1
32.3 µs ± 30 ns       48.4 µs ± 132 ns
total 100 nans 0
42.5 µs ± 45.9 ns     58.5 µs ± 140 ns
total 100 nans 1
42.9 µs ± 50.6 ns     58.9 µs ± 210 ns
total 100 nans 10
42.7 µs ± 114 ns      59.3 µs ± 259 ns
total 100 nans 50
43 µs ± 576 ns        59.3 µs ± 649 ns
total 10000 nans 0
266 µs ± 105 ns       308 µs ± 1.18 µs
total 10000 nans 100
265 µs ± 322 ns       309 µs ± 555 ns
total 10000 nans 1000
260 µs ± 301 ns       323 µs ± 6.06 µs
total 10000 nans 5000
230 µs ± 295 ns       319 µs ± 1.16 µs
```