@pllim 2022-08-02
p.s. I might have said this in a different PR of yours already, but this series of performance improvements would be nice to have a change log for.