@pllim 2024-04-23
Re-opened as requested by @maxnoe in https://github.com/astropy/astropy/pull/16326